Rename: Session을 저장하는 클래스 명을 Database에서 Storage로 변경
