+++ src/main/java/view/ProfilePage.java
@@ -10,8 +10,8 @@ import java.io.FileReader;
 import java.io.IOException;
 
 import static db.Database.findUserById;
-import static db.SessionDatabase.findAllSessionIds;
-import static db.SessionDatabase.findUserIdBySessionId;
+import static db.SessionStorage.findAllSessionIds;
+import static db.SessionStorage.findUserIdBySessionId;
 import static exception.ExceptionList.NOT_EXIST_SESSION_ID;
 import static exception.ExceptionList.NO_SESSION_ID;
 import static http.FilePath.PROFILE;
