+++ src/test/java/service/UserServiceTest.java
@@ -1,7 +1,7 @@
 package service;
 
 import db.Database;
-import db.SessionDatabase;
+import db.SessionStorage;
 import exception.NotExistUserException;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.*;
@@ -10,7 +10,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static db.Database.clear;
-import static db.SessionDatabase.*;
+import static db.SessionStorage.*;
 import static exception.ExceptionList.ALREADY_EXIST_USER;
 import static exception.ExceptionList.NOT_EXIST_USER;
 import static org.junit.jupiter.api.Assertions.*;
@@ -141,8 +141,8 @@ class UserServiceTest {
         String sessionId = findAllSessionIds().stream().findFirst().orElseGet(null);
 
         // Then
-        softAssertions.assertThat(SessionDatabase.findAllSessionIds().size())
-                .as("Database의 크기가 1이 아닙니다.\n현재 값: %d", SessionDatabase.findAllSessionIds().size())
+        softAssertions.assertThat(SessionStorage.findAllSessionIds().size())
+                .as("Database의 크기가 1이 아닙니다.\n현재 값: %d", SessionStorage.findAllSessionIds().size())
                 .isEqualTo(1);
         softAssertions.assertThat(findUserIdBySessionId(sessionId))
                 .as("userId가 'honggildong'이 아닙니다.\n현재 값: %s", findUserIdBySessionId(sessionId))
